alerts_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymongo==4.10.1
pydantic==2.10.0
requests==2.32.3
python-dotenv==1.0.1
